move bbn store logic
BBN store logic was misplaced under the branch module, which is meant to
host the general logic for branches, specifically BNP.

This change fixes that by clearing up the BBN dumper skeleton with the
store logic that subsumed it.

Besides that the BbnStoreReader was removed since there is no need for
it to exist.