mmap in reconstruction
it's incorrect to bail upon encountering an empty node. This is because
read_sequential operated in bigger chunks than a page to make reading
more efficient. However, that also means that at some point not the
whole seq read chunk would be read and `read_exact` would report
unexpected EOF error, which is funny enough is expected.

However, the documentation states that after read_exact returns
unexpected EOF the contents of the output buffer are unspecified.

I first tried to re-implement using the original approach and fix
correctness, however, the implementation turned a bit too nasty and I
realized that I can simplify by leveraging mmap + read ahead.

Here is the documentation for MADV_SEQUENTIAL

	Expect page references in sequential order.  (Hence, pages
	in the given range can be aggressively read ahead, and may
	be freed soon after they are accessed.)

Therefore theoretically this should be even more efficient than the
previous implementation.

It also seems to be more simple, but I admit that it might be because
I dealt with this particular pattern a lot of times. It could be further
simplified by sweeping the unsafes under the rug by using one of
off-the-shelf libraries providing the same functionality.